fix: prevent window from being tiled to wrong area
Using seperate functions for window moving and resizing would make windows
resize correctly but influence the delta when moving them.
For example, a big/fullscreen window could be resized correctly to a
third of its size. However, that resize would "move" the window,
influencing the delta that sould be aplied to its final position.
Using MoveResize apparently fixes this issue (it preservers the initial
position AFAIK).